finish task refactor and solve some issues

+++ core/sensors/humidity_sensors/analog_humidity_sensor.cpp
@@ -18,7 +18,7 @@ fic_error_t AnalogHumiditySensor::deinit() {
 fic_error_t AnalogHumiditySensor::measure(float &value) {
     int voltage;
     _adc->measure(voltage);
-    float humidity = 100 - 100 * voltage/_max_voltage_mv;
+    float humidity = 100 - 100 * static_cast<float>(voltage) / static_cast<float>(_max_voltage_mv);
 
     if (humidity < 0) {
         humidity = 0;
